+++ errors/error_test.go
@@ -4,52 +4,138 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"runtime/debug"
+	"strings"
 	"testing"
 )
 
-func TestStackFormatMatches(t *testing.T) {
+// fixture functions doing work to avoid inlining
+func a(i int) error {
+	if b(i + 5) && b(i + 6) {
+		return nil
+	}
+	return fmt.Errorf("not gonna happen")
+}
 
+func b(i int) bool {
+	return c(i+2) > 12
+}
+
+// panicking function!
+func c(i int) int {
+	if i > 3 {
+		panic('a')
+	}
+	return i * i
+}
+
+func TestParsePanicStack(t *testing.T) {
 	defer func() {
-		err := recover()
-		if err != 'a' {
-			t.Fatal(err)
+		err := New(recover(), 0)
+		if err.Error() != "97" {
+			t.Errorf("Received incorrect error, expected 'a' got '%s'", err.Error())
 		}
-
-		bs := [][]byte{Errorf("hi").Stack(), debug.Stack()}
-
-		// Ignore the first line (as it contains the PC of the .Stack() call)
-		bs[0] = bytes.SplitN(bs[0], []byte("\n"), 2)[1]
-		bs[1] = bytes.SplitN(bs[1], []byte("\n"), 2)[1]
-
-		if bytes.Compare(bs[0], bs[1]) != 0 {
-			t.Errorf("Stack didn't match")
-			t.Errorf("%s", bs[0])
-			t.Errorf("%s", bs[1])
+		if err.TypeName() != "*errors.errorString" {
+			t.Errorf("Error type was '%s'", err.TypeName())
 		}
+		for index, frame := range err.StackFrames() {
+			if frame.Func() == nil {
+				t.Errorf("Failed to remove nil frame %d", index)
+			}
+		}
+		expected := []StackFrame{
+			StackFrame{Name: "TestParsePanicStack.func1", File: "errors/error_test.go"},
+			StackFrame{Name: "a", File: "errors/error_test.go", LineNumber: 13},
+		}
+		assertStacksMatch(t, expected, err.StackFrames())
 	}()
 
-	a()
+	a(1)
+}
+
+func TestParseGeneratedStack(t *testing.T) {
+	err := New(fmt.Errorf("e_too_many_colander"), 0)
+	expected := []StackFrame{
+		StackFrame{Name: "TestParseGeneratedStack", File: "errors/error_test.go"},
+	}
+	if err.Error() != "e_too_many_colander" {
+		t.Errorf("Error name was '%s'", err.Error())
+	}
+	if err.TypeName() != "*errors.errorString" {
+		t.Errorf("Error type was '%s'", err.TypeName())
+	}
+	for index, frame := range err.StackFrames() {
+		if frame.Func() == nil {
+			t.Errorf("Failed to remove nil frame %d", index)
+		}
+	}
+	assertStacksMatch(t, expected, err.StackFrames())
 }
 
 func TestSkipWorks(t *testing.T) {
-
 	defer func() {
-		err := recover()
-		if err != 'a' {
-			t.Fatal(err)
+		err := New(recover(), 1)
+		if err.Error() != "97" {
+			t.Errorf("Received incorrect error, expected 'a' got '%s'", err.Error())
 		}
 
-		bs := [][]byte{New("hi", 2).Stack(), debug.Stack()}
-
-		if !bytes.HasSuffix(bs[1], bs[0]) {
-			t.Errorf("Stack didn't match")
-			t.Errorf("%s", bs[0])
-			t.Errorf("%s", bs[1])
+		for index, frame := range err.StackFrames() {
+			if frame.Name == "TestSkipWorks.func1" {
+				t.Errorf("Failed to skip frame")
+			}
+			if frame.Func() == nil {
+				t.Errorf("Failed to remove inlined frame %d", index)
+			}
 		}
+
+		expected := []StackFrame{
+			StackFrame{Name: "a", File: "errors/error_test.go", LineNumber: 13},
+		}
+
+		assertStacksMatch(t, expected, err.StackFrames())
 	}()
 
-	a()
+	a(4)
+}
+
+func checkFramesMatch(expected StackFrame, actual StackFrame) bool {
+	if actual.Name != expected.Name {
+		return false
+	}
+	// Not using exact match as it would change depending on whether
+	// the package is being tested within or outside of the $GOPATH
+	if expected.File != "" && !strings.HasSuffix(actual.File, expected.File) {
+		return false
+	}
+	if expected.Package != "" && actual.Package != expected.Package {
+		return false
+	}
+	if expected.LineNumber != 0 && actual.LineNumber != expected.LineNumber {
+		return false
+	}
+	return true
+}
+
+func assertStacksMatch(t *testing.T, expected []StackFrame, actual []StackFrame) {
+	var lastmatch int = 0
+	var matched int = 0
+	// loop over the actual stacktrace, checking off expected frames as they
+	// are found. Each one might be in the middle of the stack, but the order
+	// should remain the same.
+	for _, actualFrame := range actual {
+		for index, expectedFrame := range expected {
+			if index < lastmatch {
+				continue
+			}
+			if checkFramesMatch(expectedFrame, actualFrame) {
+				lastmatch = index
+				matched += 1
+				break
+			}
+		}
+	}
+	if matched != len(expected) {
+		t.Fatalf("failed to find matches for %d frames: '%v'\ngot: '%v'", len(expected)-matched, expected[matched:], actual)
+	}
 }
 
 type testErrorWithStackFrames struct {
@@ -122,26 +208,4 @@ func ExampleNew_skip() {
 			err = New(err, 1)
 		}
 	}()
-}
-
-func ExampleError_Stack() {
-	e := New("Oh noes!", 1)
-	fmt.Printf("Error: %s\n", e.Error())
-	fmt.Printf("Stack is %d bytes", len(e.Stack()))
-	// Output:
-	// Error: Oh noes!
-	// Stack is 589 bytes
-}
-
-func a() error {
-	b(5)
-	return nil
-}
-
-func b(i int) {
-	c()
-}
-
-func c() {
-	panic('a')
 }